pasar funciones que estaban en modelo a manager
se migra la logica que estaba en el modelo de apropiaciones al manager pendiente construir el flujo en el controlador e incluir las funciones de API financiera que correspondan a apropiacion inicial
